+++ releasenotes.md
@@ -14,7 +14,7 @@
 - Add `--list-asm` to view all supported `asm` instructions.
 - Formatting option "%h" now supports pointers.
 - Improve error on unsigned implicit conversion to signed.
-
+- Update error message for struct initialization #2286
 ### Fixes
 - mkdir/rmdir would not work properly with substring paths on non-windows platforms.
 - Hex string formatter check incorrectly rejected slices.